refactor: make init funcs to station and line
- initState.js is only for when there is nothing in local Storage.
- stattionInit, lineInit is for when the page is refreshed. They call the datas in local storage and update to HTML.
