+++ constantValues/constants.js
@@ -1,5 +1,6 @@
 const constants = {
     LOCAL_STORAGE_KEY_STATION : "station",
+    LOCAL_STORAGE_KEY_LINE : "line",
     TABLE_TARGET : "tr",
     STATION_VALUE_CHILDNODE_INDEX : 0,
     STATION_NAME_MINIMUM_LENGTH : 2,
