+++ station/stationAddRemove.js
@@ -1,5 +1,6 @@
 import {message} from "../constantValues/message.js"
 import {constants} from "../constantValues/constants.js"
+import {optionTagUpdate} from "../line/lineInit.js"
 
 const addStationPreliminaryWork = () => {
     const userInputStation = document.querySelector("#station-name-input").value;
@@ -56,21 +57,10 @@ const deleteStation = (event) => {
         
         // remove from table.
         removeTarget.remove();
+
+        // line select option update.
+        optionTagUpdate();
     }
 }
 
-const loadLocalStorageData = () => {
-    const localStorageData = localStorage.getItem(constants.LOCAL_STORAGE_KEY_STATION);
-
-    if(localStorageData) {
-        const localStorageDataToArray = JSON.parse(localStorageData);
-        
-        const stationList = document.querySelector("#station-name-table tbody");
-        stationList.innerHTML = `<tr><th>역 이름</th><th>설정</th></tr>`;
-        localStorageDataToArray.forEach(station => {
-            stationList.innerHTML += `<tr><td>${station}</td><td><button class = "station-delete-button">삭제</button></td></tr>`;
-        })
-    }
-}
-
-export {addStationPreliminaryWork, deleteStation, loadLocalStorageData, addStation};
+export {addStationPreliminaryWork, deleteStation, addStation};
